pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "streamlit>=1.49.1",
    "tensorflow>=2.20.0",
]
